requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
langchain-google-genai
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain_text_splitters
langchain-openai
langchain-anthropic
chromadb
sentence-transformers
langchain_huggingface
faiss-cpu
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
langchain-ollama